Deleting outdated stuff
Critical changes to Discord's back-end, BetterDiscord and the base themes I had provided my own aamendments to has made much of this content outdated and rather than attempting to refactor it, I decided to remove it. I'll figure out a reprise for much of this stuff later, but for now this is only going to host stuff for ChannelTabs.

if you were somehow relying on any f the deleted content for any reason, I apologize — see the previous commit from years back when i was more into this crap.