Added fav bar and settings
Added one new variables;
— CT_favBar: Adds colour to favbar (which makes up for moving the whole container up)

Modified autoohide to include fav bar and settings: those elements fade in now much like the tab bar interface buttons.

Added a second note in the theme file: theme has to be declared _second_ because of additional theming changes which are included in autohide, for when the user only wants autohide and nothing else.